Repository 인터페이스 ID 제네릭 타입 Integer로 되어있는 것들 모두 Long으로 변경

+++ src/main/java/com/project/api/bookmark/BookmarkRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
     Optional<Bookmark> findByBookmarkId(Long bookmarkId);
 
     void deleteByBookmarkId(Long bookmarkId);
