+++ src/main/java/com/project/api/test/TestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TestRepository extends JpaRepository<Test, Integer> {
+public interface TestRepository extends JpaRepository<Test, Long> {
     Optional<Test> findByTestParam2(String testParam2);
 }
